refactor(frontend): extract helpers from appointments, categories, pipe-avatar (#244)
Batch 12: Extract types, constants, and utility functions into co-located helper files.
- appointments/page.tsx: extract TabFilter, STATUS_BORDER, STATUS_DATE_BG, parseDateParts → _helpers.ts
- categories/page.tsx: extract Zod schema + CategoryFormValues type → _helpers.ts
- pipe-avatar.tsx: extract PipeMood type, color constants, MOOD_EYES, MOOD_ANIM → pipe-avatar-helpers.ts
